// LHUDatingApp/src/app/core/model/user.ts
export interface User {
    Code: string;
    UserName: string;
    Password: string;
    FullName: string;
    Dob: string;
    Phone: string;
    Email: string;
    Address: string;
    Avatar: string;
    Gender: string;
    IsMatch: number;
    Role: number;
    Token: string;

    fieldStamp1: boolean;
}